refactor(agents): unify anonymous visitor id with analytics session id
Reuse the existing persisted analytics session id (getAnalyticsSessionId,
localStorage `base44_analytics_session_id`) as the X-Base44-Anonymous-Id
value instead of minting a separate dedicated id. One identity per anonymous
visitor across analytics + agents (enables correlation), and a single source
of truth for the id.

Follow-up (deferred): harden the id generator to crypto-strength and/or
decouple agent identity from analytics — the current generateUuid is
Math.random-based, acceptable short-term but weak for a bearer-style
ownership credential.

- Drop src/utils/anon-visitor.ts; axios-client uses getAnalyticsSessionId.
- Rename test to anonymous-visitor-header.test.ts; assert the header equals
  the persisted analytics session id and is absent when authenticated.
